LOTS OF BUG FIXES, also got rid of the leave timer function

+++ external music functions/deleteNowPlaying.js
@@ -1,7 +1,5 @@
 module.exports = async function deleteNowPlaying(server) {
-    try {
+    if (server.audioStatus == 'playing') {
         await server.nowPlaying.delete();
-    } catch (err) {
-        console.log(err);
     }
 }
